Removes var.test_subdomain_name in favor of locals
I wanted to experiment with 'random' to generate the subdomain name,
rather than specify it in a variable.

I learned that in order to reference a variable in `locals`, you
must use `local` (singular).

I also made it so that my Terraform doesn't run "all the way".
Instead, with this commit you must do two steps to get this
infrastructure instantiated:

    $ terraform apply -target=aws_acm_certificate.devcert
    $ terraform apply

This is because devcert is referenced in the for_each, and with the
introduction of locals and random, I get:

    The "for_each" map includes keys derived from resource
    attributes that cannot be determined until apply, and so
    Terraform cannot determine the full set of keys that will
    identify the instances of this resource.

    When working with unknown values in for_each, it's better to
    define the map keys statically in your configuration and place
    apply-time results only in the map values.

    Alternatively, you could use the -target planning option to
    first apply only the resources that the for_each value depends
    on, and then apply a second time to fully converge.